use awesomebot to valid links in travis
Adds the travis config necessary to validate that links in
README.md are functional links.

Note that axure and education.github.com are whitelisted. Here's
why:

axure: something with the way that site does gzip make awesomebot
  think the link is down when it's not. it works just fine in a
  browser. You can see the error via
    `Faraday.get 'http://www.axure.com/edu'`
  I deemed it out of the scope of this project to debug faraday
  and just whitelisted it.
edu.github: this site is duped several times in the doc, which is
  reasonable IMO since several discounts depend on a edu github
  pack. we could allow this via an `--allow-dupes` flag, but i
  i don't think we want to allow ALL dupes. the bot should catch
  when someone adds a redundant link. But edu.github is a good
  dupe. so i whitelisted it.