Remove yellowcircle and nodehost offers
You can check both offers do not exist anymore.
1. nodehost:
https://nodehost.ca/education
Is showes 404 Page not found.
2. yellowcircle:
Over the last 8 years Yellow Circle provided a free technology platform to over 300,000 students and teachers from around the world. Our team is sincerely thankful to all of your support in the past and look forward to working with you on future projects. As of October 15, 2022 the virtual learning platform is officially retired.